Allow RAET to be PIP installable. Fixes #15
Created a new module which holds the package metadata, `__version__`,
`__author__`, `__license__`. Additionally added `__version_info__` which allows
math based version comparison.
This metadata is still available in `raet.__init__` because we import it there,
but more importantly, we can now load this metadata in the setup script without
having to load the raet module(`raet.__init__`, which in turn loads all it's
modules which in turn load the dependencies). This way, we allow pip to install
the required dependencies before actually needing then.