Bugfix: Windows exceptions may not contain 'errno'
An example of this is 'win32file.CreateFile' which contains 'winerror'
instead of 'errno'. This will fix a problem where the error handling
code would itself throw an exception. What is interesting, is that at
least in the case of 'errno.ENOENT', we can use errno values to compare
'winerror' error codes.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>